apps/german_stt_local/requirements.txt
streamlit
openai-whisper
transformers
sentencepiece
sacremoses
accelerate
